+++ src/models/properties/tests.rs
@@ -1,12 +1,10 @@
-use crate::models::{
-    properties::{DateOrDateTime, RollupPropertyValue, RollupValue},
-    PropertyValue,
-};
+use crate::models::properties::{DateOrDateTime, RollupPropertyValue, RollupValue};
+use crate::models::properties::{FormulaResultValue, PropertyValue};
 use chrono::NaiveDate;
 
 #[test]
 fn verify_date_parsing() {
-    let date = NaiveDate::from_ymd_opt(2021, 01, 02).unwrap();
+    let date = NaiveDate::from_ymd_opt(2021, 1, 2).unwrap();
     let result = serde_json::to_string(&DateOrDateTime::Date(date)).unwrap();
     let parsed: DateOrDateTime = serde_json::from_str(&result).unwrap();
     println!("{:?}", parsed);
@@ -56,4 +54,27 @@ fn parse_rollup_property() {
     {
         assert!(matches!(array[0], RollupPropertyValue::Text { .. }))
     }
+}
+
+#[test]
+fn parse_number_formula_prop() {
+    let _property: PropertyValue =
+        serde_json::from_str(include_str!("tests/formula_number_value.json")).unwrap();
+}
+
+#[test]
+fn parse_date_formula_prop() {
+    let _property: PropertyValue =
+        serde_json::from_str(include_str!("tests/formula_date_value.json")).unwrap();
+}
+
+#[test]
+fn parse_number_formula() {
+    let _value: FormulaResultValue = serde_json::from_str(
+        r#"{
+    "type": "number",
+    "number": 0
+  }"#,
+    )
+    .unwrap();
 }
